+++ AndroidStudioProject/app/src/main/java/org/gkisalatiga/plus/lib/NavigationRoutes.kt
@@ -25,6 +25,9 @@ public data class NavigationRoutes (
     val SCREEN_LITURGI: String = "nav_screen_liturgi",
     val SCREEN_VIDEO_LIST: String = "nav_screen_saren",
     val SCREEN_POSTER_VIEWER: String = "nav_screen_poster",
+    val SCREEN_AGENDA: String = "nav_screen_agenda",
+    val SCREEN_PERSEMBAHAN: String = "nav_screen_offertory",
+    val SCREEN_GALERI: String = "nav_screen_gallery",
     val SCREEN_BLANK: String = "nav_screen_blank",
 
     // The individual Composable "fragments" of each screen.