+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # GKI Salatiga+ Changelog
 
+## v0.2.1-alpha (18) --- 2024-08-04
+
+- New: Added new screens: agenda and offertory (persembahan)
+- New: Added social media call-to-action (CTA) buttons in the FragmentAbout
+- Improved: Added representative icons to main menu items
+
 ## v0.2.0-alpha (17) --- 2024-08-03
 
 - New: The main menu now has an elegant scrolling behavior